requirements.txt
PyYAML
numpy-quaternion